backend/requirements.txt
Jinja2
python-dotenv
pdflatex
python-docx
docx2txt
docxtpl
openai
